# Enable `CowData`/`String` in constexpr environment