--- a/core/math/bvh_abb.h
+++ b/core/math/bvh_abb.h
@@ -59,7 +59,6 @@ struct BVH_ABB {
 	POINT neg_max;
 
 	bool operator==(const BVH_ABB &o) const { return (min == o.min) && (neg_max == o.neg_max); }
-	bool operator!=(const BVH_ABB &o) const { return (*this == o) == false; }
 
 	void set(const POINT &_min, const POINT &_max) {
 		min = _min;