--- a/core/math/color.h
+++ b/core/math/color.h
@@ -74,9 +74,6 @@ struct [[nodiscard]] Color {
 	constexpr bool operator==(const Color &p_color) const {
 		return (r == p_color.r && g == p_color.g && b == p_color.b && a == p_color.a);
 	}
-	constexpr bool operator!=(const Color &p_color) const {
-		return (r != p_color.r || g != p_color.g || b != p_color.b || a != p_color.a);
-	}
 
 	constexpr Color operator+(const Color &p_color) const;
 	constexpr void operator+=(const Color &p_color);