--- a/core/math/plane.h
+++ b/core/math/plane.h
@@ -77,7 +77,6 @@ struct [[nodiscard]] Plane {
 	bool is_finite() const;
 
 	constexpr bool operator==(const Plane &p_plane) const;
-	constexpr bool operator!=(const Plane &p_plane) const;
 	explicit operator String() const;
 
 	Plane() = default;
@@ -127,10 +126,6 @@ Plane::Plane(const Vector3 &p_point1, const Vector3 &p_point2, const Vector3 &p_
 
 constexpr bool Plane::operator==(const Plane &p_plane) const {
 	return normal == p_plane.normal && d == p_plane.d;
-}
-
-constexpr bool Plane::operator!=(const Plane &p_plane) const {
-	return normal != p_plane.normal || d != p_plane.d;
 }
 
 template <>