--- a/core/math/rect2.h
+++ b/core/math/rect2.h
@@ -206,7 +206,7 @@ struct [[nodiscard]] Rect2 {
 	bool is_finite() const;
 
 	constexpr bool operator==(const Rect2 &p_rect) const { return position == p_rect.position && size == p_rect.size; }
-	constexpr bool operator!=(const Rect2 &p_rect) const { return position != p_rect.position || size != p_rect.size; }
+	bool operator==(const Rect2i &p_recti) const;
 
 	inline Rect2 grow(real_t p_amount) const {
 		Rect2 g = *this;