--- a/core/math/transform_2d.h
+++ b/core/math/transform_2d.h
@@ -110,7 +110,6 @@ struct [[nodiscard]] Transform2D {
 	Transform2D looking_at(const Vector2 &p_target) const;
 
 	constexpr bool operator==(const Transform2D &p_transform) const;
-	constexpr bool operator!=(const Transform2D &p_transform) const;
 
 	void operator*=(const Transform2D &p_transform);
 	Transform2D operator*(const Transform2D &p_transform) const;
@@ -157,16 +156,6 @@ constexpr bool Transform2D::operator==(const Transform2D &p_transform) const {
 	}
 
 	return true;
-}
-
-constexpr bool Transform2D::operator!=(const Transform2D &p_transform) const {
-	for (int i = 0; i < 3; i++) {
-		if (columns[i] != p_transform.columns[i]) {
-			return true;
-		}
-	}
-
-	return false;
 }
 
 constexpr void Transform2D::operator*=(real_t p_val) {