--- a/core/math/vector2.cpp
+++ b/core/math/vector2.cpp
@@ -213,3 +213,7 @@ Vector2::operator String() const {
 Vector2::operator Vector2i() const {
 	return Vector2i(x, y);
 }
+
+bool Vector2::operator==(const Vector2i &p_vector2i) const {
+	return operator==(Vector2(p_vector2i));
+}