--- a/core/math/vector3.cpp
+++ b/core/math/vector3.cpp
@@ -175,3 +175,7 @@ Vector3::operator String() const {
 Vector3::operator Vector3i() const {
 	return Vector3i(x, y, z);
 }
+
+bool Vector3::operator==(const Vector3i &p_Vector3i) const {
+	return operator==(Vector3(p_Vector3i));
+}