--- a/core/math/vector4.cpp
+++ b/core/math/vector4.cpp
@@ -225,3 +225,7 @@ static_assert(sizeof(Vector4) == 4 * sizeof(real_t));
 Vector4::operator Vector4i() const {
 	return Vector4i(x, y, z, w);
 }
+
+bool Vector4::operator==(const Vector4i &p_vector4i) const {
+	return operator==(Vector4(p_vector4i));
+}