… .github/workflows/linux_builds.yml
… core/object/object_id.h
… core/object/ref_counted.h
… core/register_core_types.cpp
… core/string/node_path.cpp
… core/string/node_path.h
… core/string/string_name.cpp
… core/string/string_name.h
… core/string/ustring.cpp
… core/string/ustring.h
… core/templates/a_hash_map.h
… core/templates/cowdata.h
… core/templates/hash_map.h
… core/templates/hash_set.h
… core/templates/interpolated_property.h
… core/templates/list.h
… core/templates/local_vector.h
… core/templates/pair.h
… core/templates/rb_map.h
… core/templates/rb_set.h
… core/templates/rid.h
… core/templates/safe_list.h
… core/templates/safe_refcount.h
… core/templates/vector.h
… core/typedefs.h
… core/variant/array.cpp
… core/variant/array.h
… core/variant/callable.cpp
… core/variant/callable.h
… core/variant/container_type_validate.h