* Depends on https://github.com/godotengine/godot/pull/100749

Enable the usage of `CowData`/`String` with string literals at compile time.

The main idea here is to construct a static `char32_t` buffer with the same layout `CowData` get at runtime. `refcount` is assigned a big enough value (currently `UINT32_MAX`) so the buffer will never be freed.

Main changes:
- `CowBuffer`: a static `char32_t` buffer with `refcount` and `size` attached at front.
- `ComptimeString`: convenient wrapper. Instead of writing `static constinit _buf = CowBuffer("str"); constexpr String s = CowData<char32_t>(buf);`, we can just use `ComptimeString<CowBuffer("str")>()`
- Change many operators of `CowData`/`String` to `constexpr`.

One possible usage of this is string literals in `GDCLASS` which I include as a showcase.

Binary Size (`production=yes target=template_release use_llvm=yes`)
- Master with #100749: 64939KB
- this PR:  64132KB (- 807KB)

Benchmarks (5 runs average)
<details>
<summary>Result (quite chaos and not very consistent)</summary>

raw data: 
[ma.jsonl.txt](https://github.com/user-attachments/files/21332907/ma.jsonl.txt)
[pr.jsonl.txt](https://github.com/user-attachments/files/21332908/pr.jsonl.txt)

```bash
                                                      pr              ma       diff    percent
Export Gltf                                       107.50          108.24   -0000.74     -00.68%
Import Fbx                                        505.12          507.40   -0002.28     -00.45%
Import Gltf                                        55.62           55.90   -0000.28     -00.50%
Import Ogg Audio                                    6.64            6.70   -0000.06     -00.89%
Import Webp Images                                806.18          805.06   +0001.12     +00.14%
Append Array                                      166.88          158.68   +0008.20     +05.17%
Bsearch                                           101.55          103.66   -0002.11     -02.04%
Fill                                               61.78           49.38   +0012.40     +25.12%
Reverse                                            25.83           25.79   +0000.05     +00.19%
Function Callable                                 317.34          313.24   +0004.10     +01.31%
Lambda Inline Callable                            373.30          386.24   -0012.94     -03.35%
Lambda Variable Callable                           90.77           89.52   +0001.25     +01.40%
Load                                               35.73           36.37   -0000.65     -01.78%
Load With Password                                 38.15           40.57   -0002.42     -05.96%
Save                                               33.81           33.46   +0000.35     +01.04%
Save With Password                                 39.15           39.02   +0000.13     +00.34%
Generate 1g Random Bytes 1k At A Time            1238.80         1247.00   -0008.20     -00.66%
Generate 1m Random Bytes 10 At A Time            2192.20         2255.20   -0063.00     -02.79%
Create                                             17.19           16.92   +0000.27     +01.60%
12 Threads Full Contention                        863.94          556.98   +0306.96     +55.11%
12 Threads Half Contention                        441.80          293.30   +0148.50     +50.63%
12 Threads Little Contention                      108.88           79.62   +0029.26     +36.76%
12 Threads Slope Contention                       468.76          328.58   +0140.18     +42.66%
2 Threads Full Contention                          62.46           55.99   +0006.47     +11.56%
2 Threads Half Contention                          43.62           41.34   +0002.28     +05.51%
2 Threads Little Contention                        28.31           27.73   +0000.58     +02.08%
4 Threads Full Contention                         184.86          163.86   +0021.00     +12.82%
4 Threads Half Contention                         107.50           93.51   +0013.99     +14.96%
4 Threads Little Contention                        42.16           38.34   +0003.81     +09.95%
4 Threads Slope Contention                        139.96          122.92   +0017.04     +13.86%
8 Threads Full Contention                         464.30          340.78   +0123.52     +36.25%
8 Threads Half Contention                         248.26          188.90   +0059.36     +31.42%
8 Threads Little Contention                        69.11           57.71   +0011.41     +19.77%
8 Threads Slope Contention                        290.88          221.20   +0069.68     +31.50%
Single                                             24.82           25.07   -0000.24     -00.97%
Global Scope Randf                                109.64          109.62   +0000.02     +00.02%
Global Scope Randf Range                          361.52          359.50   +0002.02     +00.56%
Global Scope Randfn                               852.24          855.16   -0002.92     -00.34%
Global Scope Randi                                102.90          102.68   +0000.22     +00.21%
Global Scope Randi Range                          140.54          140.06   +0000.48     +00.34%
Global Scope Randomize                            779.64          778.38   +0001.26     +00.16%
Randf                                             269.94          270.44   -0000.50     -00.18%
Randf Range                                       296.42          295.66   +0000.76     +00.26%
Randfn                                            806.48          799.56   +0006.92     +00.87%
Randi                                              84.76           86.48   -0001.73     -02.00%
Randi Range                                       113.02          112.94   +0000.08     +00.07%
Randomize                                         801.48          795.76   +0005.72     +00.72%
Emission Params 0                                 213.92          216.66   -0002.74     -01.26%
Emission Params 1                                 217.76          218.76   -0001.00     -00.46%
Emission Params 10                                252.44          254.04   -0001.60     -00.63%
Deep Tree                                         132.80          132.90   -0000.10     -00.08%
Duplicate                                        1078.00         1129.20   -0051.20     -04.53%
Fragmentation                                    1202.60         1211.80   -0009.20     -00.76%
Wide Tree                                         126.22          125.12   +0001.10     +00.88%
Fill Loop                                         241.70          242.74   -0001.04     -00.43%
Fill Method                                       105.34          105.80   -0000.46     -00.43%
Packed Color Array                                142.88          145.04   -0002.16     -01.49%
Packed Float 32 Array                             108.02          113.14   -0005.12     -04.53%
Packed Float 64 Array                             112.54          111.14   +0001.40     +01.26%
Packed Int 32 Array                                90.97           90.50   +0000.48     +00.53%
Packed Int 64 Array                                96.77           94.21   +0002.55     +02.71%
Packed String Array                               524.32          531.40   -0007.08     -01.33%
Packed Vector 2 Array                             122.28          122.18   +0000.10     +00.08%
Packed Vector 3 Array                             139.24          138.00   +0001.24     +00.90%
Typed Color Array                                 217.24          214.68   +0002.56     +01.19%
Typed Float Array                                 183.68          180.68   +0003.00     +01.66%
Typed Int Array                                   175.06          178.96   -0003.90     -02.18%
Typed String Array                                625.42          611.32   +0014.10     +02.31%
Typed Vector 2 Array                              209.70          200.24   +0009.46     +04.72%
Typed Vector 3 Array                              214.50          214.16   +0000.34     +00.16%
Untyped Color Array                               253.72          257.50   -0003.78     -01.47%
Untyped Float Array                               212.50          211.98   +0000.52     +00.25%
Untyped Int Array                                 214.48          212.42   +0002.06     +00.97%
Untyped String Array                              660.20          658.04   +0002.16     +00.33%
Untyped Vector 2 Array                            242.44          241.40   +0001.04     +00.43%
Untyped Vector 3 Array                            249.92          248.04   +0001.88     +00.76%
Binary Trees 13                                   527.26          529.46   -0002.20     -00.42%
Binary Trees 15                                  2441.00         2470.20   -0029.20     -01.18%
For Loop Add                                        8.11            7.85   +0000.25     +03.21%
For Loop Call                                      62.61           60.69   +0001.92     +03.16%
Lambda Call                                        46.62           45.11   +0001.51     +03.35%
Mandelbrot Set                                   2430.80         2298.20   +0132.60     +05.77%
Merkle Trees 13                                  1635.80         1637.60   -0001.80     -00.11%
Merkle Trees 15                                  7879.20         7941.80   -0062.60     -00.79%
Nbody 1 000 000                                  5383.20         5161.60   +0221.60     +04.29%
Nbody 500 000                                    2700.60         2589.40   +0111.20     +04.29%
Spectral Norm 100                                  50.07           49.70   +0000.37     +00.75%
Spectral Norm 1000                               4604.60         4645.80   -0041.20     -00.89%
Spectral Norm 500                                1170.60         1160.80   +0009.80     +00.84%
Md 5 Buffer Empty                                 206.84          206.24   +0000.60     +00.29%
Md 5 Buffer Non Empty                             674.36          678.36   -0004.00     -00.59%
Md 5 Text Empty                                   172.76          173.10   -0000.34     -00.20%
Md 5 Text Non Empty                               643.18          641.78   +0001.40     +00.22%
Sha 1 Buffer Empty                                190.72          189.10   +0001.62     +00.86%
Sha 1 Buffer Non Empty                            590.30          595.74   -0005.44     -00.91%
Sha 1 Text Empty                                  161.20          153.54   +0007.66     +04.99%
Sha 1 Text Non Empty                              563.54          552.86   +0010.68     +01.93%
Sha 256 Buffer Empty                              264.68          267.08   -0002.40     -00.90%
Sha 256 Buffer Non Empty                          894.52          900.94   -0006.42     -00.71%
Sha 256 Text Empty                                235.52          238.78   -0003.26     -01.37%
Sha 256 Text Non Empty                            866.80          887.90   -0021.10     -02.38%
Complex Variable Concatenate                     1522.40         1549.60   -0027.20     -01.76%
Complex Variable Method                          2647.60         2839.40   -0191.80     -06.75%
Complex Variable Percent                         2644.60         2663.60   -0019.00     -00.71%
No Op Constant Method                             140.18          139.24   +0000.94     +00.68%
Simple Constant Concatenate                         4.61            4.58   +0000.03     +00.68%
Simple Constant Method                            554.14          554.48   -0000.34     -00.06%
Simple Constant Method Constant Dict              325.68          330.22   -0004.54     -01.37%
Simple Constant Percent                             4.64            4.61   +0000.03     +00.56%
Simple Variable Concatenate                       156.78          152.62   +0004.16     +02.73%
Simple Variable Method                            556.10          556.86   -0000.76     -00.14%
Simple Variable Percent                           421.62          426.90   -0005.28     -01.24%
Begins With                                        11.15           11.08   +0000.07     +00.65%
Bigrams                                           722.12          938.22   -0216.10     -23.03%
Capitalize                                        968.02         1018.50   -0050.48     -04.96%
Casecmp To                                          8.82            8.78   +0000.04     +00.48%
Contains                                            9.56            9.55   +0000.01     +00.11%
Contains Gdscript In                                4.64            4.62   +0000.02     +00.42%
Count                                              55.79           52.50   +0003.30     +06.28%
Countn                                            246.92          249.50   -0002.58     -01.03%
Ends With                                          11.02           10.89   +0000.13     +01.21%
Find                                               35.22           35.16   +0000.07     +00.19%
Findn                                              65.64           64.80   +0000.83     +01.29%
Get Slice                                          70.78           71.15   -0000.36     -00.51%
Get Slice Count                                    21.44           21.60   -0000.16     -00.74%
Humanize Size                                     701.94          710.34   -0008.40     -01.18%
Insert                                             62.57           61.45   +0001.12     +01.82%
Is Valid Filename                                  23.11           22.92   +0000.19     +00.81%
Lpad                                              138.40          137.68   +0000.72     +00.52%
Naturalnocasecmp To                                12.27           12.52   -0000.24     -01.95%
Nocasecmp To                                       37.30           37.44   -0000.13     -00.35%
Pad Decimals                                      439.60          438.46   +0001.14     +00.26%
Pad Decimals Pre Constructed                       72.41           72.88   -0000.47     -00.65%
Pad Zeros                                         251.22          251.72   -0000.50     -00.20%
Pad Zeros Pre Constructed                         164.02          166.06   -0002.04     -01.23%
Rfind                                              40.54           42.24   -0001.70     -04.03%
Rfindn                                            183.64          179.74   +0003.90     +02.17%
Rpad                                              126.02          127.42   -0001.40     -01.10%
Rsplit                                            401.38          402.78   -0001.40     -00.35%
Similarity                                         22.96           22.68   +0000.28     +01.22%
Simplify Path                                    1189.40         1392.20   -0202.80     -14.57%
Split                                             383.92          379.06   +0004.86     +01.28%
Split Floats                                      320.14          317.98   +0002.16     +00.68%
Substr                                             60.69           59.76   +0000.93     +01.56%
To Camel Case                                     627.48          660.94   -0033.46     -05.06%
To Lower                                          195.54          213.44   -0017.90     -08.39%
To Pascal Case                                    909.52          949.96   -0040.44     -04.26%
To Snake Case                                     869.56          887.76   -0018.20     -02.05%
To Utf 16 Buffer                                  147.04          146.64   +0000.40     +00.27%
To Utf 32 Buffer                                   98.14           99.23   -0001.09     -01.10%
To Utf 8 Buffer                                   142.52          144.56   -0002.04     -01.41%
To Wchar Buffer                                   145.28          148.20   -0002.92     -01.97%
Uri Decode                                        489.94          500.06   -0010.12     -02.02%
Uri Encode                                        419.00          444.62   -0025.62     -05.76%
Validate Filename                                 304.42          307.32   -0002.90     -00.94%
Validate Node Name                                 72.48           73.65   -0001.17     -01.59%
Xml Escape                                        427.68          439.72   -0012.04     -02.74%
Xml Unescape                                       76.40           75.42   +0000.98     +01.29%
Both Clean And Simplify                          5957.80         6026.00   -0068.20     -01.13%
Only Clean                                       5953.20         5939.00   +0014.20     +00.24%
Only Simplify                                      14.68           16.06   -0001.38     -08.57%
Quickest                                           13.34           17.42   -0004.07     -23.38%
Delaunay 2d                                       119.98          114.24   +0005.74     +05.02%
Delaunay 3d                                       611.70          599.80   +0011.90     +01.98%
Parse 20 Complex Expressions Cold                 291.12          289.98   +0001.14     +00.39%
Parse 20 Complex Expressions With Cached Parse          286.76          286.24   +0000.52     +00.18%
Parse 20 Complex Expressions With Cached Parse Then Execute          292.96          291.74   +0001.22     +00.42%
Cellular                                         1152.80         1168.20   -0015.40     -01.32%
Perlin                                            406.08          405.02   +0001.06     +00.26%
Simplex                                           408.02          411.92   -0003.90     -00.95%
Simplex Smooth                                    500.98          509.06   -0008.08     -01.59%
Value                                             403.38          401.82   +0001.56     +00.39%
Value Cubic                                       710.92          710.00   +0000.92     +00.13%
Triangulate                                        94.96           94.78   +0000.18     +00.19%
Astar 5000 Times 5000 Points                      424.40          422.18   +0002.22     +00.53%
Astar 1000 Times 1000 Points                      564.24          563.44   +0000.80     +00.14%
1000 Moving Agents                                  0.02            0.02   -0000.00     -03.70%
Navigation 10 000 Random Paths                     10.90           11.66   -0000.76     -06.56%
Add Children With Same Name                        50.78           49.72   +0001.07     +02.14%
Add Children Without Name                          56.67           52.97   +0003.70     +06.98%
Delete Children In Order                           10.99           10.96   +0000.03     +00.24%
Delete Children Random Order                       13.13           13.21   -0000.08     -00.61%
Delete Children Reverse Order                       6.68            6.44   +0000.24     +03.76%
Get Node                                           30.69           30.06   +0000.62     +02.07%
Move Children                                     709.00          746.72   -0037.72     -05.05%
Activate 1024 Viewports                           227.02          222.78   +0004.24     +01.90%
Activate 256 Viewports                            120.64          120.58   +0000.06     +00.05%
Activate 64 Viewports                               6.80            6.72   +0000.08     +01.16%
Create 1024 Viewports                             128.40          122.40   +0006.00     +04.90%
Create 256 Viewports                               50.93           50.49   +0000.44     +00.88%
Create 64 Viewports                                 3.51            3.26   +0000.25     +07.74%
Sum of all tests                                97692.78        97476.29   +0216.49     +00.22%
```
…