@Ivorforce 2025-07-29
Hi, this is an interesting proposition that we've been discussing a few times!

I have also made an attempt at implementing it, which you may want to have a look at: https://github.com/Ivorforce/godot/tree/constexpr-string

I don't exactly remember the outcome, but I believe the main problem I encountered was that users needed to use a custom macro (`SString`) to make use of `constexpr` strings. Ideally, the `constexpr` constructor would be automatically chosen for `char[]` values, although I don't know if that's possible (mind you, I haven't looked at your implementation yet!).