@YYF233333 2025-07-29
> I don't exactly remember the outcome, but I believe the main problem I encountered was that users needed to use a custom macro (`SString`) to make use of `constexpr` strings. Ideally, the `constexpr` constructor would be automatically chosen for `char[]` values, although I don't know if that's possible (mind you, I haven't looked at your implementation yet!).

Yes, I encountered the same issue. I believe the key here is that we must somehow write `static constexpr str_storage = "literals";`, because currently we cannot dynamically allocate memory at compile time, so the user must provide external memory to store the string. I provide a template struct `ComptimeString` which is basicly identical to your macro. Haven't find a way to automatically convert string literals to `constexpr String` yet as it is not possible now to get the space for `refcount` and `size` in functions.